+++ BackEnd/src/main/java/BackEnd/Form/ProductForm/BrandForm/BrandDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class BrandDTO {
 
-    private Byte brandId;
+    private Integer brandId;
 
     private String brandName;
