+++ BackEnd/src/main/java/BackEnd/Form/ProductForm/ShoeTypeForm/ShoeTypeDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ShoeTypeDTO {
 
-    private Byte shoeTypeId;
+    private  Integer shoeTypeId;
 
     private String shoeTypeName;
