+++ BackEnd/src/main/java/BackEnd/Repository/ProductRepository/IBrandRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.List;
 
-public interface IBrandRepository extends JpaRepository<Brand, Byte> , JpaSpecificationExecutor<Brand> {
+public interface IBrandRepository extends JpaRepository<Brand, Integer> , JpaSpecificationExecutor<Brand> {
 
 }
