+++ BackEnd/src/main/java/BackEnd/Repository/ProductRepository/IColorRepository.java
@@ -8,9 +8,9 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface IColorRepository extends JpaRepository<Color, Byte>, JpaSpecificationExecutor<Color> {
+public interface IColorRepository extends JpaRepository<Color,  Integer>, JpaSpecificationExecutor<Color> {
 
     @Query("SELECT c FROM Color c JOIN ShoeColor sc ON c.id = sc.id.colorId WHERE sc.id.shoeId = :shoeId")
-    List<Color> findColorsByShoeId(@Param("shoeId") Short shoeId);
+    List<Color> findColorsByShoeId(@Param("shoeId")  Integer shoeId);
 
 }
