+++ BackEnd/src/main/java/BackEnd/Service/ProductService/Brand/BrandService.java
@@ -44,7 +44,7 @@ public class BrandService implements IBrandService {
     }
 
     @Override
-    public Brand getBrandById(Byte id) {
+    public Brand getBrandById( Integer id) {
         return IBrandRepository.findById(id).orElse(null);
     }
 
@@ -94,7 +94,7 @@ public class BrandService implements IBrandService {
         2. Xóa ảnh trong Server
         3. Xóa "Brand" khỏi database
      */
-    public void deleteBrand(Byte brandId) {
+    public void deleteBrand( Integer brandId) {
         shoeService.updateDefaultBrandOfShoes(brandId);
         Brand oldBrand = getBrandById(brandId);
         ImageService.deleteImage(ImageService.brandLogoPath, oldBrand.getLogo());
